automation/pyproject.toml
[project]
name = "account-mall-automation"
version = "0.1.0"
description = "Local Apple ID automation (status test, change password, change region) via CLI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "botasaurus>=4.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
]

[project.scripts]
apple-automation = "cli:app"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
